fix(AI书源): 修复温度赋值 Float 转 Double 类型错误

+++ app/src/main/java/io/legado/app/ui/book/source/ai/AiSourceGenerateActivity.kt
@@ -399,7 +399,7 @@ class AiSourceGenerateActivity :
             .setTitle("高级设置")
             .setView(dialogBinding.root)
             .setPositiveButton("确定") { _, _ ->
-                viewModel.temperature = dialogBinding.seekTemp.progress / 20f
+                viewModel.temperature = (dialogBinding.seekTemp.progress / 20f).toDouble()
                 dialogBinding.etHtmlLimit.text?.toString()?.toIntOrNull()?.let {
                     viewModel.promptHtmlLimit = it
                 }
